% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_log.R
\name{address.trace}
\alias{address.trace}
\title{identify the address trace with error}
\usage{
address.trace(err.node)
}
\arguments{
\item{err.node}{error node}
}
\value{
stack if stack trace found else the aux
}
\description{
identify the address trace with error
}
