% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_log_read.R
\name{binary_log_read}
\alias{binary_log_read}
\title{user_error_display}
\usage{
binary_log_read(logfile)
}
\arguments{
\item{logfile}{path to the logfile}
}
\value{
count.dt list with error messgae,inputs which caused the error,
  corresponding line and datavariable name
}
\description{
user_error_display
}
\author{
Akhila Chowdary Kolla
}
