% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_setup.R
\name{ci_setup}
\alias{ci_setup}
\title{RcppDeepState-action GitHub workflow setup}
\usage{
ci_setup(
  repository = "./",
  event = "pull_request",
  fail_ci_if_error = FALSE,
  location = "/",
  seed = "-1",
  time_limit = "2",
  max_inputs = "3",
  comment = FALSE
)
}
\arguments{
\item{repository}{path to the repository root folder}

\item{event}{the event that triggers the workflow to run, accepted values 
are pull_request and push}

\item{fail_ci_if_error}{specify if CI pipeline should fail when 
RcppDeepState finds errors}

\item{location}{location of the package if not in the root of the repository}

\item{seed}{seed used for deterministic fuzz testing and reproduce the 
analysis results}

\item{time_limit}{fuzzing phase's duration in seconds}

\item{max_inputs}{number of inputs generated in the fuzzing phase to analyze}

\item{comment}{print the analysis results as a comment in pull request}
}
\description{
RcppDeepState-action GitHub workflow setup
}
