% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_make.R
\name{create_specific_makefile}
\alias{create_specific_makefile}
\title{creates makefiles for above created testharness in package}
\usage{
create_specific_makefile(package, fun_name)
}
\arguments{
\item{package}{to the RcppExports file}

\item{fun_name}{name of function to get makefile}
}
\description{
creates makefiles for above created testharness in package
}
