% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-objects.R
\name{create_testpkgs_objects}
\alias{create_testpkgs_objects}
\title{creates testharness for given functions in package}
\usage{
create_testpkgs_objects()
}
\description{
creates testharness for given functions in package
}
