% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary.R
\name{deep_harness_analyze_one}
\alias{deep_harness_analyze_one}
\title{analyze the binary file}
\usage{
deep_harness_analyze_one(path)
}
\arguments{
\item{path}{to test}
}
\value{
returns a list of all the param values of the arguments of function
}
\description{
analyze the binary file
}
