% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcppTestharness.R
\name{deep_harness_compile_run}
\alias{deep_harness_compile_run}
\title{compiles the code for created testharness in package}
\usage{
deep_harness_compile_run(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\description{
compiles the code for created testharness in package
}
