% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_log_binary.R
\name{deep_harness_log_binary}
\alias{deep_harness_log_binary}
\title{analyze the binary file}
\usage{
deep_harness_log_binary(path)
}
\arguments{
\item{path}{to test}
}
\description{
analyze the binary file
}
