% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tests.R
\name{deep_harness_save_crash_tests}
\alias{deep_harness_save_crash_tests}
\title{compile the code and save crashu=ing tests}
\usage{
deep_harness_save_crash_tests(package_name)
}
\arguments{
\item{package_name}{for the RcppExports file}
}
\description{
compile the code and save crashu=ing tests
}
