% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcppTestharness.R
\name{deep_harness_save_passing_tests}
\alias{deep_harness_save_passing_tests}
\title{compile the code and save passing tests}
\usage{
deep_harness_save_passing_tests(package_name)
}
\arguments{
\item{package_name}{for the RcppExports file}
}
\description{
compile the code and save passing tests
}
