% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_fuzzer_builds.R
\name{deepstate_afl_make}
\alias{deepstate_afl_make}
\title{creates afl fuzzer for given functions in package}
\usage{
deepstate_afl_make()
}
\description{
creates afl fuzzer for given functions in package
}
