% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allchecks.R
\name{deepstate_allchecks}
\alias{deepstate_allchecks}
\title{analyze the asan and valgrind checks}
\usage{
deepstate_allchecks(path)
}
\arguments{
\item{path}{to the package we are working on}
}
\description{
analyze the asan and valgrind checks
}
