% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary_one.R
\name{deepstate_analyze_fun}
\alias{deepstate_analyze_fun}
\title{analyze the binary for one function}
\usage{
deepstate_analyze_fun(file.path)

deepstate_analyze_fun(file.path)
}
\arguments{
\item{file.path}{file to analyze}

\item{fun_path}{function path to run}

\item{max_inputs}{input arguments}
}
\description{
analyze the binary for one function

analyze the binary for one function
}
