% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary_one.R
\name{deepstate_analyze_one}
\alias{deepstate_analyze_one}
\title{analyze the binary for one function}
\usage{
deepstate_analyze_one(fun_path)
}
\arguments{
\item{fun_path}{function path to run}
}
\description{
analyze the binary for one function
}
