% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_log_read.R
\name{deepstate_binary_log_read}
\alias{deepstate_binary_log_read}
\title{deepstate_binary_log_read}
\usage{
deepstate_binary_log_read(logfile)
}
\arguments{
\item{logfile}{path to the logfile}
}
\value{
count.dt list with error messgae,inputs which caused the error,
  corresponding line and datavariable name
}
\description{
deepstate_binary_log_read
}
\author{
Akhila Chowdary Kolla
}
