% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable_fun.R
\name{deepstate_compile_checks_fun}
\alias{deepstate_compile_checks_fun}
\title{Checks Testharness compilation and execution}
\usage{
deepstate_compile_checks_fun(
  package_path,
  function_name,
  continue_on_missing = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{package_path}{path to the testpackage}

\item{function_name}{function name in the package}

\item{continue_on_missing}{if set to FALSE, terminates the execution if 
asserts are missing}

\item{verbose}{used to deliver more in depth information}
}
\description{
This function compiles and runs the checks test harness.
The checks test harness contains user defined assertions
}
