% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_compile_one.R
\name{deepstate_compile_fun}
\alias{deepstate_compile_fun}
\title{compile the  for one function}
\usage{
deepstate_compile_fun(fun_path, sep = "infun")
}
\arguments{
\item{fun_path}{function path to compile}

\item{sep}{type of harness to compile}
}
\description{
compile the  for one function
}
