% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable_fun.R
\name{deepstate_compile_generate_fun}
\alias{deepstate_compile_generate_fun}
\title{Generation test harness compilation and execution}
\usage{
deepstate_compile_generate_fun(
  package_path,
  function_name,
  continue_on_missing = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{package_path}{path to the tested package}

\item{function_name}{function name in the package}

\item{continue_on_missing}{if set to FALSE, terminates the execution if 
at leas one range is missing}

\item{verbose}{used to deliver more in depth information}
}
\description{
This function compiles and runs the generation test harness.
The generation test harness contains user defined ranges
}
