% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcpppackageTestHarness.R
\name{deepstate_create_makefile}
\alias{deepstate_create_makefile}
\title{creates makefiles for above created testharness in package}
\usage{
deepstate_create_makefile(package, fun_name)
}
\arguments{
\item{package}{to the RcppExports file*}

\item{fun_name}{name of function to get makefile}
}
\description{
creates makefiles for above created testharness in package
}
