% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkgs.R
\name{deepstate_datatype_check}
\alias{deepstate_datatype_check}
\title{deepstate_datatype_check}
\usage{
deepstate_datatype_check(params.list)
}
\arguments{
\item{params.list}{to the package test file}
}
\description{
deepstate_datatype_check
}
