% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdsp.R
\name{deepstate_display}
\alias{deepstate_display}
\title{deepstate_display}
\usage{
deepstate_display(logfile)
}
\arguments{
\item{logfile}{path to the logfile}
}
\value{
count.dt list with error messgae
}
\description{
deepstate_display
}
\author{
Akhila Chowdary Kolla
}
