% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleoutput.R
\name{deepstate_displays}
\alias{deepstate_displays}
\title{deepstate_displays}
\usage{
deepstate_displays(logfile)
}
\arguments{
\item{logfile}{path to the logfile}
}
\description{
deepstate_displays
}
\author{
Akhila Chowdary Kolla
}
