% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable_fun.R
\name{deepstate_editable_fun}
\alias{deepstate_editable_fun}
\title{creates testharness for given function in package}
\usage{
deepstate_editable_fun(package_name, function_name)
}
\arguments{
\item{package_name}{to the RcppExports file}

\item{function_name}{of the package}
}
\description{
creates testharness for given function in package
}
