% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfunction.R
\name{deepstate_fun_create}
\alias{deepstate_fun_create}
\title{creates testharness for given function in package}
\usage{
deepstate_fun_create(package_name, function_name, sep = "infun")
}
\arguments{
\item{package_name}{to the RcppExports file}

\item{function_name}{of the package}

\item{sep}{param to decide infun or outfun}
}
\description{
creates testharness for given function in package
}
