% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary.R
\name{deepstate_fuzz_fun_analyze}
\alias{deepstate_fuzz_fun_analyze}
\title{analyze the binary file}
\usage{
deepstate_fuzz_fun_analyze(test_function, seed = -1, time.limit.seconds)
}
\arguments{
\item{test_function}{path of the test function}

\item{seed}{input seed to pass on the executable}

\item{time.limit.seconds}{duration to run the code}
}
\description{
analyze the binary file
}
