% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_details.R
\name{deepstate_get_function_body}
\alias{deepstate_get_function_body}
\title{gets function body}
\usage{
deepstate_get_function_body(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\value{
function.list list with function names and arguments
}
\description{
gets function body
}
\examples{
deepstate_get_function_body(system.file("testpkgs/testSAN", package = "RcppDeepState")) 
}
