% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pkgs.R
\name{deepstate_get_mismatched_datatypes}
\alias{deepstate_get_mismatched_datatypes}
\title{deepstate_get_mismatched_datatypes}
\usage{
deepstate_get_mismatched_datatypes(params.list)
}
\arguments{
\item{params.list}{to the package test file}
}
\description{
deepstate_get_mismatched_datatypes
}
