% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_details.R
\name{deepstate_get_package_details}
\alias{deepstate_get_package_details}
\title{gets package details}
\usage{
deepstate_get_package_details(path)
}
\arguments{
\item{path}{to the RcppExports file}
}
\value{
function list with function names and arguments
}
\description{
gets package details
}
