% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_details.R
\name{deepstate_get_prototype_calls}
\alias{deepstate_get_prototype_calls}
\title{Prototypes}
\usage{
deepstate_get_prototype_calls(package_path)
}
\arguments{
\item{package_path}{to the RcppExports file}
}
\value{
prototypes list with function prototypes
}
\description{
Prototypes
}
