% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary.R
\name{deepstate_harness_analyze_pkg}
\alias{deepstate_harness_analyze_pkg}
\title{analyze the binary file}
\usage{
deepstate_harness_analyze_pkg(path, testfiles = "all", max_inputs = "all")
}
\arguments{
\item{path}{to test}

\item{testfiles}{no of test files to test}

\item{max_inputs}{no of bin files to analyze}
}
\value{
returns a list of all the param values of the arguments of function
}
\description{
analyze the binary file
}
