% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_names.R
\name{deepstate_harness_files}
\alias{deepstate_harness_files}
\title{creates list of files and path for given functions in package}
\usage{
deepstate_harness_files(path)
}
\arguments{
\item{path}{to the package test file}
}
\value{
makefiles.list makefile list
}
\description{
creates list of files and path for given functions in package
}
