% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tests.R
\name{deepstate_harness_save_passing_tests}
\alias{deepstate_harness_save_passing_tests}
\title{compile the code and save passing tests}
\usage{
deepstate_harness_save_passing_tests(package_name)
}
\arguments{
\item{package_name}{for the RcppExports file}
}
\description{
compile the code and save passing tests
}
