% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_names.R
\name{deepstate_list_bin_directory}
\alias{deepstate_list_bin_directory}
\title{creates list of log files}
\usage{
deepstate_list_bin_directory(path)
}
\arguments{
\item{path}{to the package test file}
}
\value{
bindir.list returns binary folder paths
}
\description{
creates list of log files
}
