% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_names.R
\name{deepstate_list_log_files}
\alias{deepstate_list_log_files}
\title{creates list of log files}
\usage{
deepstate_list_log_files(path)
}
\arguments{
\item{path}{to the package test file}
}
\value{
logfiles.list logfile path list
}
\description{
creates list of log files
}
