% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_names.R
\name{deepstate_list_package_args}
\alias{deepstate_list_package_args}
\title{creates list of log files}
\usage{
deepstate_list_package_args(path)
}
\arguments{
\item{path}{to the package test file}
}
\value{
list.args.i returns args path list
}
\description{
creates list of log files
}
