% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_run.R
\name{deepstate_make_run}
\alias{deepstate_make_run}
\title{Deepstate Static library}
\usage{
deepstate_make_run(verbose = getOption("verbose"))
}
\arguments{
\item{verbose}{used to deliver more in depth information}
}
\description{
Creates a deepstate static library for the user in the ~/.RcppDeepState directory.
}
