% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcppTestharness.R
\name{deepstate_pkg_create}
\alias{deepstate_pkg_create}
\title{creates testharness for given functions in package}
\usage{
deepstate_pkg_create(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\description{
creates testharness for given functions in package
}
