% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_log.R
\name{deepstate_read_valgrind_xml}
\alias{deepstate_read_valgrind_xml}
\title{return datatable for parse xml data}
\usage{
deepstate_read_valgrind_xml(xml.file)
}
\arguments{
\item{xml.file}{the input log file path}
}
\value{
line.num.dt list with error messgae,line number where the error ocuured
}
\description{
return datatable for parse xml data
}
