% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_testing.R
\name{deepstate_tests_fuzz}
\alias{deepstate_tests_fuzz}
\title{deepstate_tests_fuzz}
\usage{
deepstate_tests_fuzz(package_path)
}
\arguments{
\item{package_path}{path to the package to test}
}
\description{
deepstate_tests_fuzz
}
\author{
Akhila Chowdary Kolla
}
