% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asan_error_log.R
\name{deepstate_user_asan_error_display}
\alias{deepstate_user_asan_error_display}
\title{analyze the binary file}
\usage{
deepstate_user_asan_error_display(path)
}
\arguments{
\item{path}{to test}
}
\value{
returns a list of all the param values of log files
}
\description{
analyze the binary file
}
