% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorrcode.R
\name{deepstate_user_error_display}
\alias{deepstate_user_error_display}
\title{user_error_display}
\usage{
deepstate_user_error_display(logfile)
}
\arguments{
\item{logfile}{path to the logfile}
}
\value{
count.dt list with error messgae,inputs which caused the error,
  corresponding line and datavariable name
}
\description{
user_error_display
}
\author{
Akhila Chowdary Kolla
}
