% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_log.R
\name{deepstate_xmlog}
\alias{deepstate_xmlog}
\title{return datatable for parse xml data}
\usage{
deepstate_xmlog(log)
}
\arguments{
\item{log}{the input log file path}
}
\description{
return datatable for parse xml data
}
