% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcppTestharness.R
\name{get_function_body}
\alias{get_function_body}
\title{gets function body}
\usage{
get_function_body(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\value{
function.list list with function names and arguments
}
\description{
gets function body
}
