% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcppTestharness.R
\name{get_package_details}
\alias{get_package_details}
\title{gets package details}
\usage{
get_package_details(path)
}
\arguments{
\item{path}{to the RcppExports file}
}
\value{
function list with function names and arguments
}
\description{
gets package details
}
