% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_details.R
\name{get_package_name}
\alias{get_package_name}
\title{get_package_name}
\usage{
get_package_name(package_path)
}
\arguments{
\item{package_path}{location of the package in the filesystem}
}
\description{
gets the name of a package from its DESCRIPTION file
}
