% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRcppTestharness.R
\name{get_prototype_calls}
\alias{get_prototype_calls}
\title{gets prototype calls}
\usage{
get_prototype_calls(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\value{
prototypes list with function prototype
}
\description{
gets prototype calls
}
