% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_names.R
\name{harness_files}
\alias{harness_files}
\title{creates list of files and path for given functions in package}
\usage{
harness_files(path)
}
\arguments{
\item{path}{to the package test file}
}
\description{
creates list of files and path for given functions in package
}
