% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary.R
\name{inputs.table}
\alias{inputs.table}
\title{analyze the binary file}
\usage{
inputs.table(inputs.column)
}
\arguments{
\item{inputs.column}{inputs column of result table}
}
\description{
analyze the binary file
}
