% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary.R
\name{issues.table}
\alias{issues.table}
\title{analyze the binary file}
\usage{
issues.table(logtable.list)
}
\arguments{
\item{logtable.list}{logtable  column of result table}
}
\description{
analyze the binary file
}
