% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_create_run.R
\name{package_deep_harness_compile_run}
\alias{package_deep_harness_compile_run}
\title{compiles the code for created testharness in package}
\usage{
package_deep_harness_compile_run(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\description{
compiles the code for created testharness in package
}
