% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_create_run.R
\name{package_deepstate_pkg_create}
\alias{package_deepstate_pkg_create}
\title{creates testharness for given functions in package}
\usage{
package_deepstate_pkg_create(package_name)
}
\arguments{
\item{package_name}{to the RcppExports file}
}
\description{
creates testharness for given functions in package
}
