\name{RcppDeepState-package}
\alias{RcppDeepState-package}
\alias{RcppDeepState}
\docType{package}
\title{\packageTitle{RcppDeepState}}
\description{\packageDescription{RcppDeepState}}
\section{Package Content}{\packageIndices{RcppDeepState}}
\author{\packageAuthor{RcppDeepState}}
\section{Maintainer}{\packageMaintainer{RcppDeepState}}
\keyword{package}
