% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary.R
\name{rcppdeepstate_compile_run_analyze}
\alias{rcppdeepstate_compile_run_analyze}
\title{analyze the binary file}
\usage{
rcppdeepstate_compile_run_analyze(test_function, seed, time.limit.seconds)
}
\arguments{
\item{test_function}{path of the test package}

\item{seed}{input seed to pass on the executable}

\item{time.limit.seconds}{duration to run the code}
}
\description{
analyze the binary file
}
