% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_log.R
\name{stack.trace}
\alias{stack.trace}
\title{identify the stack trace with error}
\usage{
\method{stack}{trace}(s.node, src.dir.xpath)
}
\arguments{
\item{s.node}{stack frame}

\item{src.dir.xpath}{gives path to the src file}
}
\value{
trace with the error
}
\description{
identify the stack trace with error
}
