% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_binary_one.R
\name{deepstate_analyze_file}
\alias{deepstate_analyze_file}
\title{analyze the binary for one function}
\usage{
deepstate_analyze_file(file.path)
}
\arguments{
\item{file.path}{file to analyze}
}
\description{
analyze the binary for one function
}
