% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_harness_compile.R
\name{deepstate_compile_fun}
\alias{deepstate_compile_fun}
\title{Harness compilation for the function}
\usage{
deepstate_compile_fun(fun_path, sep = "infun")
}
\arguments{
\item{fun_path}{path of the function to compile}

\item{sep}{default to infun}
}
\description{
Compiles the function-specific testharness in the package.
}
\examples{
pkg_path <- file.path("./RcppDeepState/testpkgs/testSAN")
deepstate_pkg_create(pkg_path)
path <-  file.path("./RcppDeepState/testpkgs/testSAN/inst/testfiles/rcpp_read_out_of_bound")
compiled.harness <- deepstate_compile_fun(path)
print(compiled.harness)
}
