% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_harness_compile.R
\name{deepstate_compile_fun}
\alias{deepstate_compile_fun}
\title{Harness compilation for the function}
\usage{
deepstate_compile_fun(fun_path, sep = "infun")
}
\arguments{
\item{fun_path}{path of the function to compile}

\item{sep}{default to infun}
}
\description{
Compiles the function-specific testharness in the package.
}
\examples{
path <- system.file("testpkgs/testSAN", package = "RcppDeepState")
compiled.harness <- deepstate_compile_fun(path)
print(compiled.harness)
}
