% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable_fun.R
\name{deepstate_compile_generate_fun}
\alias{deepstate_compile_generate_fun}
\title{compiles generate harness for given function in package}
\usage{
deepstate_compile_generate_fun(package_name, function_name)
}
\arguments{
\item{package_name}{to the RcppExports file}

\item{function_name}{of the package}
}
\description{
compiles generate harness for given function in package
}
