% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable_fun.R
\name{deepstate_compile_generate_fun}
\alias{deepstate_compile_generate_fun}
\title{Generation Testharness compilation}
\usage{
deepstate_compile_generate_fun(package_path, function_name)
}
\arguments{
\item{package_path}{path to the testpackage}

\item{function_name}{function name in the package}
}
\description{
This function compiles the generation testharness.
}
