% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_makefile_create.R
\name{deepstate_create_makefile}
\alias{deepstate_create_makefile}
\title{Harness Makefile Generation}
\usage{
deepstate_create_makefile(package, fun_name)
}
\arguments{
\item{package}{path to the test package}

\item{fun_name}{name of the function}
}
\description{
This function generates makefile for the provided function specific TestHarness
}
