% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_harness_create.R
\name{deepstate_fun_create}
\alias{deepstate_fun_create}
\title{TestHarness for the function}
\usage{
deepstate_fun_create(package_path, function_name, sep = "infun")
}
\arguments{
\item{package_path}{to the test package}

\item{function_name}{from the test package}

\item{sep}{infun default}
}
\value{
The TestHarness file that is generated
}
\description{
The function creates Testharness for the provided function name.
}
\examples{
path <- system.file("testpkgs/testSAN", package = "RcppDeepState")
function_name <- "rcpp_read_out_of_bound"
function.harness <- deepstate_fun_create(path,function_name)
print(function.harness)
}
