% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_compile_one.R
\name{deepstate_fuzz_fun}
\alias{deepstate_fuzz_fun}
\title{analyze the binary file}
\usage{
deepstate_fuzz_fun(fun_path, seed = -1, time.limit.seconds = -1)
}
\arguments{
\item{fun_path}{path of the test package function}

\item{seed}{input seed to pass on the executable}

\item{time.limit.seconds}{duration to run the code}
}
\description{
analyze the binary file
}
