% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_compile_one.R
\name{deepstate_fuzz_fun}
\alias{deepstate_fuzz_fun}
\title{compile the  for one function}
\usage{
deepstate_fuzz_fun(fun_path)
}
\arguments{
\item{fun_path}{function path to run}
}
\description{
compile the  for one function
}
