% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_harness_fuzz.R
\name{deepstate_fuzz_fun}
\alias{deepstate_fuzz_fun}
\title{Harness execution for the function}
\usage{
deepstate_fuzz_fun(
  package_path,
  fun_name,
  seed = -1,
  time.limit.seconds = 2,
  sep = "infun"
)
}
\arguments{
\item{package_path}{path to the testpackage}

\item{fun_name}{name of the function to compile}

\item{seed}{input seed value to pass on the executable}

\item{time.limit.seconds}{duration to run the executable}

\item{sep}{default to infun}
}
\value{
The executed function.
}
\description{
Executes the function-specific testharness in the package.
}
\examples{
package_path <- system.file("testpkgs/testSAN", package = "RcppDeepState")
fun_name <- "rcpp_read_out_of_bound"
compiled.harness <- deepstate_fuzz_fun(package_path,fun_name)
print(compiled.harness)
#Runs the harness specified number of seconds
compiled.harness.timer <- deepstate_fuzz_fun(package_path,fun_name,time.limit.seconds=6)
print(compiled.harness.timer)
}
