% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_details.R
\name{deepstate_get_function_body}
\alias{deepstate_get_function_body}
\title{Function Details}
\usage{
deepstate_get_function_body(package_path)
}
\arguments{
\item{package_path}{path to the test package}
}
\value{
A list with function names and arguments
}
\description{
Function Details
}
\examples{
deepstate_get_function_body(system.file("testpkgs/testSAN", package = "RcppDeepState")) 
}
