% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_details.R
\name{deepstate_get_package_details}
\alias{deepstate_get_package_details}
\title{Package Details}
\usage{
deepstate_get_package_details(path)
}
\arguments{
\item{path}{to the package with RcppExports file}
}
\value{
A list with all the relevant data in the RcppExrpots
}
\description{
Ths function takes the path to the test package and captures the argument specific data.
}
