% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_harness_compile.R
\name{deepstate_harness_compile_run}
\alias{deepstate_harness_compile_run}
\title{Harness compilation for the package}
\usage{
deepstate_harness_compile_run(package_path, time.limit.seconds = 5)
}
\arguments{
\item{package_path}{to the test package}

\item{time.limit.seconds}{duration to run the testharness, defaulted to 5 seconds}
}
\value{
A character vector of compiled functions.
}
\description{
Compiles all the generated function-specific testharness in the package.
}
\examples{
path <- system.file("testpkgs/testSAN", package = "RcppDeepState")
compiled.harness.list <- deepstate_harness_compile_run(path)
print(compiled.harness.list)
}
