% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_harness_create.R
\name{deepstate_pkg_create}
\alias{deepstate_pkg_create}
\title{TestHarness for the package}
\usage{
deepstate_pkg_create(package_path)
}
\arguments{
\item{package_path}{path to the test package}
}
\value{
A character vector of TestHarness files that are generated
}
\description{
Creates Testharness for all the functions in the package that you want to test
using RcppDeepState.
}
\examples{
path <- system.file("testpkgs/testSAN", package = "RcppDeepState")
harness.list <- deepstate_pkg_create(path)
print(harness.list)
}
