% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_log.R
\name{deepstate_xmlog}
\alias{deepstate_xmlog}
\title{analyze the binary for one function}
\usage{
deepstate_xmlog(log)
}
\arguments{
\item{fun_path}{function path to run}

\item{max_inputs}{input arguments}
}
\description{
analyze the binary for one function
}
